backend/requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-deepgram
livekit-plugins-elevenlabs
livekit-plugins-silero
python-dotenv
llama-index
llama-index-vector-stores-chroma
llama-index-embeddings-openai
chromadb
pymupdf
youtube-transcript-api
trafilatura
beautifulsoup4
python-docx
tavily-python
fastapi
uvicorn
